pyproject.toml
[tool.poetry]
name = "medical-appointment"
version = "0.1.0"
description = ""
authors = ["Only-bottle <bmlee@nota.ai>"]
readme = "README.md"
packages = [{include = "medical_appointment"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.3"
djangorestframework = "^3.14.0"
black = "^23.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
